+++ CHANGELOG.md
@@ -8,6 +8,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 ### Fixed
 
+## 2022-12-08: v0.10.2
+### Fixed
+- Missing export of AppAgentWebsocket
+
+## 2022-12-08: v0.10.1
+### Fixed
+- Switch event emitter package from "events" to "Emittery".
+
 ## 2022-12-07: v0.10.0
 ### Added
 - Interface AppAgentClient, along with a class that implements that interface AppAgentWebsocket as part of work to unify the holochain and holo client apis. Happ devs will now be able to write the majority of their client code using the AppAgentClient interface, completely agnostic as to whether they're in a pure holochain or holo context.